Dockerized the backend
