+++ backend/py_neuro/.dockerignore
@@ -1,0 +1,9 @@
+__pycache__
+*.pyc
+.git
+.env
+.venv
+.pytest_cache
+.coverage
+temp_uploads/
+*.log
